+++ resources/config/config.yml
@@ -2,7 +2,7 @@ version: "0.8.0"
 
 corpusPath: "./resources/dataset/dataseer/corpus"
 templatePath: "./resources/dataset/dataseer/crfpp-templates/dataseer.template"
-grobidHome: "../grobid-home"
+grobidHome: "../../grobid/grobid-home"
 tmpPath: "tmp/"
 
 # path to Pub2TEI repository as available at https://github.com/kermitt2/Pub2TEI
