+++ src/test/java/org/grobid/core/engines/DataseerClassifierIntegrationTest.java
@@ -3,13 +3,11 @@ package org.grobid.core.engines;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import org.apache.commons.io.IOUtils;
-import org.grobid.core.data.Dataset;
 import org.grobid.core.main.GrobidHomeFinder;
 import org.grobid.core.main.LibraryLoader;
 import org.grobid.core.utilities.DatastetConfiguration;
 import org.grobid.core.utilities.GrobidConfig.ModelParameters;
 import org.grobid.core.utilities.GrobidProperties;
-import org.grobid.service.configuration.DatastetServiceConfiguration;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Ignore;
@@ -25,7 +23,7 @@ import java.util.List;
  * @author Patrice
  */
 @Ignore
-public class DatasetParserTest {
+public class DataseerClassifierIntegrationTest {
     private static DatastetConfiguration configuration;
 
     @BeforeClass
@@ -65,7 +63,7 @@ public class DatasetParserTest {
     }
 
     @Test
-    public void testDatasetParserText() throws Exception {
+    public void testDataseerBinaryClassifierText() throws Exception {
         String text = IOUtils.toString(this.getClass().getResourceAsStream("/texts.txt"), StandardCharsets.UTF_8.toString());
         String[] textPieces = text.split("\n");
         List<String> texts = new ArrayList<>();
@@ -74,17 +72,7 @@ public class DatasetParserTest {
             //System.out.println(text);
             texts.add(text);
         }
-
-        List<List<Dataset>> results = DatasetParser.getInstance(configuration).processingStrings(texts, false);
-        StringBuilder json = new StringBuilder();
-
-        int i = 0;
-        for(List<Dataset> result : results) {
-            for(Dataset dataset : result) {
-                json.append(dataset.toJson());
-                json.append("\n");
-            }
-        }
+        String json = DataseerClassifier.getInstance().classify(texts);
         System.out.println(json);
     }
